Fixed api_client not Roundrobining.
Added a .cleari() to the API connection pool before each test, forcing a new connection. this also indicates that the python client is kinda broken, it should open enough contact to get an effective round-robin in the pool. further investigation is needed, this however is good enough for this tool.